refactor: extract shared UI layout
